Reorganize namespaces + pollute less with includes bench: 4713175
